Refactor form manipulation tools into firedrake.formmanipulation
Pull ExtractSubBlock up one level, and reimplement form splitting for
the form compiler in terms of it.
