+++ firedrake/tsfc_interface.py
@@ -9,14 +9,9 @@ import gzip
 import os
 import zlib
 import tempfile
-import numpy
 import collections
 
-from ufl import Form, as_vector
-from ufl.corealg.map_dag import MultiFunction
-from ufl.algorithms.map_integrands import map_integrand_dags
-from ufl.constantvalue import Zero
-from firedrake.ufl_expr import Argument
+from ufl import Form
 
 from tsfc import compile_form as tsfc_compile_form
 
@@ -26,105 +21,9 @@ from pyop2.mpi import COMM_WORLD, dup_comm, free_comm
 
 from coffee.base import Invert
 
+from firedrake.formmanipulation import split_form
+
 from firedrake.parameters import parameters as default_parameters
-
-
-SplitForm = collections.namedtuple("SplitForm", ["indices", "form"])
-
-
-class FormSplitter(MultiFunction):
-
-    """Split a form in a list of subtrees for each component of the
-    mixed space it is built on.  See :meth:`split` for a usage
-    description."""
-
-    def split(self, form):
-        """Split the form.
-
-        :arg form: the form to split.
-
-        This is a no-op if none of the arguments in the form are
-        defined on :class:`~.MixedFunctionSpace`\s.
-
-        The return-value is a tuple for which each entry is.
-
-        .. code-block:: python
-
-           (argument_indices, form)
-
-        Where ``argument_indices`` is a tuple indicating which part of
-        the mixed space the form belongs to, it has length equal to
-        the number of arguments in the form.  Hence functionals have
-        a 0-tuple, 1-forms have a 1-tuple and 2-forms a 2-tuple
-        of indices.
-
-        For example, consider the following code:
-
-        .. code-block:: python
-
-            V = FunctionSpace(m, 'CG', 1)
-            W = V*V*V
-            u, v, w = TrialFunctions(W)
-            p, q, r = TestFunctions(W)
-            a = q*u*dx + p*w*dx
-
-        Then splitting the form returns a tuple of two forms.
-
-        .. code-block:: python
-
-           ((0, 2), w*p*dx),
-            (1, 0), q*u*dx))
-
-        """
-        args = form.arguments()
-        if all(len(a.function_space()) == 1 for a in args):
-            # No mixed spaces, just return the form directly.
-            idx = tuple([0]*len(form.arguments()))
-            return (SplitForm(indices=idx, form=form), )
-        forms = []
-        # How many subspaces do we have for each argument?
-        shape = tuple(len(a.function_space()) for a in args)
-        # Walk over all the indices of the spaces
-        for idx in numpy.ndindex(shape):
-            # Which subspace are we currently interested in?
-            self.idx = dict(enumerate(idx))
-            # Cache for the arguments we construct
-            self._args = {}
-            # Visit the form
-            f = map_integrand_dags(self, form)
-            # Zero-simplification may result in an empty form, only
-            # collect those that are non-zero.
-            if len(f.integrals()) > 0:
-                forms.append(SplitForm(indices=idx, form=f))
-        return tuple(forms)
-
-    expr = MultiFunction.reuse_if_untouched
-
-    def multi_index(self, o):
-        return o
-
-    def argument(self, o):
-        V = o.function_space()
-        if len(V) == 1:
-            # Not on a mixed space, just return ourselves.
-            return o
-        # Already seen this argument, return the cached version.
-        if o in self._args:
-            return self._args[o]
-        args = []
-        for i, V_i in enumerate(V.split()):
-            # Walk over the subspaces and build a vector that is zero
-            # where the argument does not match the one we're looking
-            # for and is just the non-mixed argument when we do want
-            # it.
-            a = Argument(V_i, o.number(), part=o.part())
-            indices = numpy.ndindex(a.ufl_shape)
-            if self.idx[o.number()] == i:
-                args += [a[j] for j in indices]
-            else:
-                args += [Zero() for j in indices]
-        self._args[o] = as_vector(args)
-        return self._args[o]
 
 
 KernelInfo = collections.namedtuple("KernelInfo",
@@ -279,7 +178,7 @@ def compile_form(form, name, parameters=None, inverse=False):
     # A map from all form coefficients to their number.
     coefficient_numbers = dict((c, n)
                                for (n, c) in enumerate(form.coefficients()))
-    for idx, f in FormSplitter().split(form):
+    for idx, f in split_form(form):
         # Map local coefficient numbers (as seen inside the
         # compiler) to the global coefficient numbers
         number_map = dict((n, coefficient_numbers[c])
